requirements.txt
Django == 3.2.19
django-cors-headers 
djangorestframework
isodate
requests
django-filter
